+++ blocklets/ai-studio/api/src/routes/resource.ts
@@ -2,6 +2,7 @@
 import { mkdir, writeFile } from 'fs/promises';
 import path from 'path';
 
+import downloadImage from '@api/libs/download-logo';
 import logger from '@api/libs/logger';
 import component from '@blocklet/sdk/lib/component';
 import { Router } from 'express';
@@ -9,7 +10,6 @@ import Joi from 'joi';
 import cloneDeep from 'lodash/cloneDeep';
 import groupBy from 'lodash/groupBy';
 import uniq from 'lodash/uniq';
-import uniqBy from 'lodash/uniqBy';
 import { stringify } from 'yaml';
 
 import { ensurePromptsEditor } from '../libs/security';
@@ -110,15 +110,30 @@ export function resourceRoutes(router: Router) {
       const { key, value } = item;
 
       const folderPath = path.join(resourceDir, key);
-      await mkdir(folderPath, { recursive: true });
 
       for (const projectId of value) {
-        const project = await Project.findOne({ where: { _id: projectId } });
-        const assistants = await getAssistantsOfRepository({ projectId, ref: project?.gitDefaultBranch! });
-
         await mkdir(path.join(folderPath, projectId), { recursive: true });
 
-        const result = stringify({ assistants: uniqBy(assistants, 'id'), project: project && project.dataValues });
+        const project = await Project.findOne({ where: { _id: projectId } });
+        const assistants = await Promise.all(
+          (await getAssistantsOfRepository({ projectId, ref: project?.gitDefaultBranch! })).map(async (i) => {
+            const logo = i.release?.logo;
+            if (!logo) return i;
+
+            const logoFilename = `${i.id}-release-logo.png`;
+            const logoPath = path.join(folderPath, projectId, logoFilename);
+
+            try {
+              await downloadImage(logo, logoPath);
+            } catch (error) {
+              logger.error('failed to download assistant logo', { error, logo });
+            }
+
+            return { ...i, release: { ...i.release, logo: logoFilename } };
+          })
+        );
+
+        const result = stringify({ assistants, project: project && project.dataValues });
 
         // TODO 兼容老的 ai assistant, 晚些去掉
         await writeFile(path.join(folderPath, `${projectId}.yaml`), result);