+++ blocklets/ai-studio/src/pages/project/save-button.tsx
@@ -19,7 +19,7 @@ import {
 } from '@mui/material';
 import { useKeyPress } from 'ahooks';
 import { bindDialog, bindTrigger, usePopupState } from 'material-ui-popup-state/hooks';
-import { useCallback, useEffect } from 'react';
+import { useCallback, useEffect, useRef } from 'react';
 import { Controller, useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
 import { joinURL } from 'ufo';
@@ -57,6 +57,10 @@ export default function SaveButton({ projectId, gitRef }: { projectId: string; g
     if (dialogState.isOpen) form.reset();
     if (branches.includes(gitRef)) {
       form.setValue('branch', gitRef);
+    }
+
+    if (!dialogState.isOpen) {
+      savePromise.current?.resolve({ saved: false });
     }
   }, [dialogState.isOpen]);
 
@@ -96,9 +100,11 @@ export default function SaveButton({ projectId, gitRef }: { projectId: string; g
         run();
         setProjectCurrentBranch(projectId, branch);
         if (branch !== gitRef) navigate(joinURL('..', branch), { replace: true });
+        savePromise.current?.resolve?.({ saved: true });
       } catch (error) {
         form.reset(input);
         Toast.error(getErrorMessage(error));
+        savePromise.current?.reject?.(error);
         throw error;
       }
     },
@@ -127,8 +133,15 @@ export default function SaveButton({ projectId, gitRef }: { projectId: string; g
     }
   );
 
+  const savePromise = useRef<{ resolve: (result: { saved?: boolean }) => void; reject: (error: Error) => void }>();
+
   useEffect(() => {
-    saveButtonState.getState().setSaveHandler(() => dialogState.open());
+    saveButtonState.getState().setSaveHandler(() => {
+      return new Promise<{ saved?: boolean } | undefined>((resolve, reject) => {
+        savePromise.current = { resolve, reject };
+        dialogState.open();
+      });
+    });
     return () => saveButtonState.getState().setSaveHandler(undefined);
   }, [dialogState.open]);